pyproject.toml
[tool.poetry]
name = "django-blog"
version = "0.1.0"
description = ""
authors = ["Andrea Maestri <16215183+andreamaestri@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
Django = "^5.1.3"
dj-database-url = "^2.3.0"
gunicorn = "^23.0.0"
psycopg2-binary = "^2.9.10"
python-dotenv = "^1.0.1"
whitenoise = "^6.8.2"
django-tinymce = "^4.1.0"
django-filebrowser = "^4.0.3"
django-admin-interface = "^0.29.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
